Keep workflow concurrency parsing together
Amp-Thread-ID: https://ampcode.com/threads/T-01a060a9-0bce-75a8-a2cc-17bfb8b4d63c
